backend/pyproject.toml
[project]
name = "cortex-backend"
version = "0.1.0"
description = "Backend for Outlook Cortex (EmailOps Edition)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "sqlalchemy>=2.0.0",
    "pgvector>=0.2.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "email-validator>=2.1.0",
    "python-dotenv>=1.0.0",
    "langgraph>=0.0.10",
    "google-cloud-aiplatform>=1.38.0",
    "openai>=1.0.0",
    "tenacity>=8.2.0",
    "structlog>=23.1.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "unstructured>=0.10.0",
    "spacy>=3.7.0",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "boto3>=1.28.0",
    "python-dateutil>=2.8.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-requests",
    "types-PyYAML",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true